完善前端基本功能，完善后端基本功能

+++ course-selection-system-back/src/course/course.controller.ts
@@ -1,15 +1,20 @@
-import { Controller, Get, Post, Body, Put, Param, Delete, ParseIntPipe } from '@nestjs/common';
+import { Controller, Get, Post, Body, Put, Param, Delete, ParseIntPipe, Req, Query } from '@nestjs/common';
 import { CourseService } from './course.service';
 import { CreateCourseDto } from './dto/create-course.dto';
 import { UpdateCourseDto } from './dto/update-course.dto';
 import { Course } from './entities/course.entity';
+import { ForbiddenException } from '@nestjs/common';
 
 @Controller('courses')
 export class CourseController {
   constructor(private readonly courseService: CourseService) {}
 
   @Post()
-  create(@Body() createCourseDto: CreateCourseDto): Promise<Course> {
+  create(@Body() createCourseDto: CreateCourseDto,@Req() req): Promise<Course> {
+    const user=req.user;
+    if(!user||user.usertype=='student'){
+      throw new Error('Unauthorized');
+    }
     return this.courseService.create(createCourseDto);
   }
 
@@ -24,12 +29,25 @@ export class CourseController {
   }
 
   @Put(':id')
-  update(@Param('id',ParseIntPipe) id: number, @Body() updateCourseDto: UpdateCourseDto): Promise<Course|null> {
+  async update(@Param('id',ParseIntPipe) id: number, @Body() updateCourseDto: UpdateCourseDto,@Req() req): Promise<Course|null> {
+    const user=req.user;
+    if(!user||user.usertype=='student'){
+      throw new ForbiddenException('Unauthorized');
+    }
     return this.courseService.update(id, updateCourseDto);
   }
 
   @Delete(':id')
-  remove(@Param('id',ParseIntPipe) id: number): Promise<void> {
+  async remove(@Param('id',ParseIntPipe) id: number,@Req() req): Promise<void> {
+    const user=req.user;
+    if(!user||user.usertype=='student'){
+      throw new ForbiddenException('Unauthorized');
+    }
     return this.courseService.remove(id);
+  }
+
+  @Get('search/:name')
+  async searchByName(@Param('name') name: string): Promise<Course[]> {
+    return this.courseService.searchByName(name);
   }
 }
